+++ external/fv3fit/tests/training/test_train.py
@@ -38,7 +38,7 @@ NON_REGRESSION_TRAINING_TYPES = ["min_max_novelty_detector", "ocsvm_novelty_dete
 # cannot be used in generic tests below
 # you must write a separate file that specializes each of the tests
 # for models in this list
-SPECIAL_TRAINING_TYPES = ["min_max_novelty_detector", "ocsvm_novelty_detector"]
+SPECIAL_TRAINING_TYPES = ["graph", "min_max_novelty_detector", "ocsvm_novelty_detector"]
 
 
 # automatically test on every registered training class
